feat: Add camera shake functionality
This commit adds camera shake functionality to the `Camera` class. It introduces a `shake` method that takes a strength and duration parameter to initiate the shaking effect. The shaking effect is implemented using random offsets within the specified strength range. The `Timer` class is also added to handle the duration of the shaking effect.